+++ R/data_2020.R
@@ -16,7 +16,7 @@
 #'   \item{afromexican}{afromexican population according to the Censo 2020}
 #'   \item{indigenous_language}{Number of persons who speak an indigenous language according to the Censo 2020}
 #' }
-#' @name df_mxstate
+#' @name df_mxstate_2020
 #' @docType data
 #' @references Population estimates taken from the \href{https://www.inegi.org.mx/programas/ccpv/2020/default.html#Tabulados}{Censo 2020.}
 #'
@@ -49,7 +49,7 @@
 #'   \item{long}{longitude of the localidad cabecera of each municipio}
 #'   \item{lat}{latitude of the localidad cabecera of each municipio}
 #' }
-#' @name df_mxmunicipio
+#' @name df_mxmunicipio_2020
 #' @docType data
 #' @references Population estimates taken from the \href{https://www.inegi.org.mx/programas/ccpv/2020/default.html#Tabulados}{Censo 2020}.
 #' The latitude and longitude of the localidad cabecera of each municipio come from the